Added QuadTree into project.
It can be used to quickly query for objects that are near a point.
And can search for type(s), sort and run a filter function.

Nodes added are updated in the quad tree very server step with their X
and Y however if this is not fast enough you could update the node
manually by calling its update function.