Moved Npc into generic/Npc.js
Added eyes to inspect things nicer.
Please do a npm install or npm upgrade.

Fixed a problem in vmscript where it was failing to handle boolean type
dependencies. Such as infos.Npc.Loaded it would continue if it was false
because it was only checking if it was not undefined.

Started coding a function to turn restruct definitions into cheat engine
structs or other kind of structure definitions.

Updated action.js and Character and Npc structures.
We can now select nodes from the quad tree based on what the player is
attacking/targeting.

Time to get monsters and items back in :D.