Fixed a bug with combineRate & Random
I did not debugged / tested this code change. If theres a mistake once again I am sorry :)

When looking to the code I figured it out something was wrong. What would happen if the item was combine 0 ? Wouldn't enter in the first if, it would throw the Error Invalide combine value. So I did a slight change on the if.
Another thing I have been confused was the input.item1 getting the combine instead of the item2 .. I guess the item1 is the left square and item 2 is the right looking at this image (http://i.imgur.com/qtJVpNH.png) . So I guess it should be the item2 getting the combine and not the item1. :)

If my change is wrong please reply why , thanks :)

Regards