Race Conditions with translation loading bleh.
#3
If game had not loaded the data yet then the pointer to data would be 0.

There might be other issues later on with data being only half loaded.
Must try to find out ItemInfo Object knows if its fully loaded or not i
think there is a boolean for Loaded.