Health, Chi, Damage, Defense and others
* Health, chi, damage, defense, all resists, clan elemental damage, etc.
is now being correctly counted and updated to the state,  using the
items that you are wearing (pets are not coded yet).
* Update when you equip or de-equip the item.
* Still some issues with dodge and hit rating with items only above 69%
enchant. The items that counts in this issue are boots and gloves. #95
* HP and Chi regeneration every 2 seconds for 400 every tick. (This one
still needs to be worked out of how much per tick we should regenerate,
so that maybe will include a level range tick or percentage of max
health/chi). #98
* statInfo is renamed to infos (client.character.infos).
* Transactions with other characters are now being re-examined and
recoded.